+++ src/Cart/Models/Cart.php
@@ -123,6 +123,11 @@ class Cart extends Model implements CartContract
         return $this->items->sum('total');
     }
 
+    public function itemsTotal(): float
+    {
+        return $this->items->sum('total');
+    }
+
     /**
      * The cart's user relationship
      *